requirements.txt
PyYAML>=6.0
requests>=2.32.0
playwright>=1.46.0
